Fix menu_id and nexted exception issue and create enum classes
