+++ src/main/java/menuManegment/demo/menu/enums/Content.java
@@ -1,0 +1,9 @@
+package menuManegment.demo.menu.enums;
+
+public enum Content {
+        childmenu,
+        content,
+        dynamic,
+        parentcat;
+    // = ["Child Menu Item", "Content", "Dynamic Content Tab", "Sub-Categories"];
+};
