+++ src/main/java/menuManegment/demo/menu/enums/Link.java
@@ -1,0 +1,7 @@
+package menuManegment.demo.menu.enums;
+
+public enum Link {
+        custom_link,
+        category_link;
+     // = ["Custom Link", "Category Link"];
+};
