+++ src/main/java/menuManegment/demo/menu/enums/Status.java
@@ -1,0 +1,6 @@
+package menuManegment.demo.menu.enums;
+
+public enum Status {
+        DISABLE,
+        ENABLE;
+};
