+++ src/main/java/menuManegment/demo/menu/enums/Target.java
@@ -1,0 +1,10 @@
+package menuManegment.demo.menu.enums;
+
+public enum Target {
+        _blank,
+        _self,
+        _parent,
+        _top;
+     // = ["Load in a new window", "Load in the same frame as it was clicked",
+    // "Load in the parent frameset", "Load in the full body of the window"];
+};
