+++ src/main/java/menuManegment/demo/menu/mapper/MegaMenuItemMapper.java
@@ -1,9 +1,13 @@
 package menuManegment.demo.menu.mapper;
 
+import menuManegment.demo.menu.entity.MegaMenu;
 import menuManegment.demo.menu.entity.MegaMenuItem;
 import menuManegment.demo.menu.model.MegaMenuItemModel;
+import menuManegment.demo.menu.model.MegaMenuModel;
 import org.mapstruct.*;
 import org.springframework.stereotype.Component;
+
+import java.util.List;
 
 @Component
 @Mapper(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
@@ -12,5 +16,13 @@ public interface MegaMenuItemMapper extends GenericMapper<MegaMenuItem, MegaMenu
     @Override
     @Mapping(target = "menu.menuItems", ignore = true)
     MegaMenuItemModel toModel(MegaMenuItem entity);
+
+    @Override
+    @Mapping(target = "menu.menuItems", ignore = true)
+    List<MegaMenuItemModel> toModels(List<MegaMenuItem> entities);
+
+//    @Override
+//    @Mapping(target = "menuItems.menu", ignore = true)
+//    List<MegaMenuItem> toEntities(List<MegaMenuItemModel> models);
 
 }
