+++ src/main/java/menuManegment/demo/menu/mapper/MegaMenuMapper.java
@@ -5,6 +5,8 @@ import menuManegment.demo.menu.model.MegaMenuModel;
 import org.mapstruct.*;
 import org.springframework.stereotype.Component;
 
+import java.util.List;
+
 @Component
 @Mapper(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
 public interface MegaMenuMapper extends GenericMapper<MegaMenu, MegaMenuModel> {
@@ -12,4 +14,12 @@ public interface MegaMenuMapper extends GenericMapper<MegaMenu, MegaMenuModel> {
     @Override
     @Mapping(target = "menuItems.menu", ignore = true)
     MegaMenuModel toModel(MegaMenu entity);
+
+    @Override
+    @Mapping(target = "menuItems.menu", ignore = true)
+    List<MegaMenuModel> toModels(List<MegaMenu> entities);
+
+//    @Override
+//    @Mapping(target = "menuItems.menu", ignore = true)
+//    List<MegaMenu> toEntities(List<MegaMenuModel> models);
 }
