+++ src/main/java/menuManegment/demo/menu/service/CRUD.java
@@ -20,8 +20,6 @@ public interface CRUD<E extends Loadable<?>, M extends ModelLoadable<?>> {
 
     List<M> retrieveAll(List<Integer> ids);
 
-    //List<E> updateStatus(String value, List<M> models);
-
     M retrieve(Loadable<Integer> id);
 
     List<M> retrieves();
